refact: nostr types
