+++ apps/mobile/store/accountBuilder.ts
@@ -5,12 +5,8 @@ import { create } from 'zustand'
 import { PIN_KEY } from '@/config/auth'
 import { getItem } from '@/storage/encrypted'
 import { type EntropyType } from '@/types/logic/entropy'
-import {
-  type Account,
-  type DM,
-  type Key,
-  type Secret
-} from '@/types/models/Account'
+import { type Account, type Key, type Secret } from '@/types/models/Account'
+import { type NostrDM } from '@/types/models/Nostr'
 import { aesDecrypt, aesEncrypt } from '@/utils/crypto'
 
 type AccountBuilderState = {
@@ -272,7 +268,7 @@ const useAccountBuilderStore = create<
         deviceNpub: '',
         deviceNsec: '',
         trustedMemberDevices: [],
-        dms: [] as DM[],
+        dms: [] as NostrDM[],
         lastUpdated: new Date(),
         syncStart: new Date()
       }
